+++ functions.php
@@ -152,6 +152,8 @@ if (!function_exists('onepress_setup')) :
 		 */
 		add_theme_support('editor-styles');
 		add_theme_support('align-wide');
+		
+		add_theme_support( 'wp-block-styles' );
 
 		/*
 		 * This theme styles the visual editor to resemble the theme style.
@@ -339,6 +341,9 @@ function onepress_scripts()
 			$onepress_js_settings['is_home'] = 1;
 		}
 	}
+	
+	$onepress_js_settings['parallax_speed'] = 0.5;
+	$onepress_js_settings =  apply_filters( 'onepress_js_settings', $onepress_js_settings );
 	wp_localize_script('onepress-theme', 'onepress_js_settings', $onepress_js_settings);
 }
 add_action('wp_enqueue_scripts', 'onepress_scripts');